+++ src/App/Bulma.fs
@@ -41,3 +41,62 @@ let styleHelpers = [
 
 let withBulmaHelpers s =
     s @ styleHelpers
+
+
+//TODO - Move into library
+module FontAwesome =
+    let fa name = Html.i [ class' ("fa fa-" + name) ]
+
+type ColumnOptions() =
+    member _.is (n:int) = class' ("is-" + string n)
+    member _.tabletIs (n:int) = class' (sprintf "is-%d-tablet" n)
+    member _.desktopIs (n:int) = class' (sprintf "is-%d-desktop" n)
+    member _.widescreenIs (n:int) = class' (sprintf "is-%d-widescreen" n)
+
+type ColumnsOptions() =
+    member _.isCentered = class' "is-centered"
+
+type HeroOptions() =
+    member _.isPrimary = class' "is-primary"
+    member _.isInfo = class' "is-info"
+    member _.isFullheight = class' "is-fullheight"
+
+type ControlOptions() =
+    member _.hasIconsLeft = class' "has-icons-left"
+
+type IconOptions() =
+    member _.isSmall = class' "is-small"
+    member _.isLeft = class' "is-left"
+
+type ButtonOptions() =
+    member _.isSuccess = class' "is-success"
+    member _.isCancel = class' "is-cancel"
+
+type BulmaEngine() =
+    member _.heroBody (props : NodeFactory list) = Html.div ([ class' "hero-body" ] @ props)
+    member _.hero (props : NodeFactory list) = Html.section ([ class' "hero" ] @ props)
+    member _.container (props : NodeFactory list) = Html.div ([ class' "container" ] @ props)
+    member _.section (props : NodeFactory list) = Html.div ([ class' "section" ] @ props)
+    member _.columns (props : NodeFactory list) = Html.div ([ class' "columns" ] @ props)
+    member _.column (props : NodeFactory list) = Html.div ([ class' "column" ] @ props)
+    member _.formBox (props : NodeFactory list) = Html.form ([ class' "box" ] @ props)
+    member _.box (props : NodeFactory list) = Html.div ([ class' "box" ] @ props)
+    member _.field (props : NodeFactory list) = Html.div ([ class' "field" ] @ props)
+    member _.label (props : NodeFactory list) = Html.label ([ class' "label" ] @ props)
+    member _.label (label:string) = Html.label ([ class' "label" ] @ [ text label ])
+    member _.labelFor (target:string) (label:string) = Html.label [ class' "label"; for' target; text label ]
+    member _.button (props : NodeFactory list) = Html.button ([ class' "button" ] @ props)
+    member _.control (props : NodeFactory list) = Html.div ([ class' "control" ] @ props)
+    member _.email (props : NodeFactory list) = Html.input ([ class' "input"; type' "email" ] @ props)
+    member _.checkbox (props : NodeFactory list) = Html.input ([ type' "checkbox" ] @ props)
+    member _.password (props : NodeFactory list) = Html.input ([ class' "input"; type' "password" ] @ props)
+    member _.icon (props : NodeFactory list) = Html.span ([ class' "icon"; type' "email" ] @ props)
+    member x.labelCheckbox (label:string) checkboxProps= Html.label [ class' "checkbox"; x.checkbox checkboxProps; text label ]
+
+let bulma = BulmaEngine()
+let hero = HeroOptions()
+let columns = ColumnsOptions()
+let column = ColumnOptions()
+let control = ControlOptions()
+let icon = IconOptions()
+let button = ButtonOptions()